basis: switch to the 8-56 representation scheme

+++ basis/linhat/closed_test.go
@@ -9,22 +9,22 @@ import (
 func TestClosedCompute(t *testing.T) {
 	basis := NewClosed(1)
 
-	compute := func(level, order uint32, point float64) float64 {
+	compute := func(level, order uint64, point float64) float64 {
 		return basis.Compute([]uint64{compose(level, order)}, []float64{point})
 	}
 
-	points := []float64{0, 0.25, 0.5, 0.75, 1}
+	points := []float64{0.0, 0.25, 0.5, 0.75, 1.0}
 
 	cases := []struct {
-		level  uint32
-		order  uint32
+		level  uint64
+		order  uint64
 		values []float64
 	}{
-		{0, 0, []float64{1, 1.0, 1, 1.0, 1}},
-		{1, 0, []float64{1, 0.5, 0, 0.0, 0}},
-		{1, 2, []float64{0, 0.0, 0, 0.5, 1}},
-		{2, 1, []float64{0, 1.0, 0, 0.0, 0}},
-		{2, 3, []float64{0, 0.0, 0, 1.0, 0}},
+		{0, 0, []float64{1.0, 1.0, 1.0, 1.0, 1.0}},
+		{1, 0, []float64{1.0, 0.5, 0.0, 0.0, 0.0}},
+		{1, 2, []float64{0.0, 0.0, 0.0, 0.5, 1.0}},
+		{2, 1, []float64{0.0, 1.0, 0.0, 0.0, 0.0}},
+		{2, 3, []float64{0.0, 0.0, 0.0, 1.0, 0.0}},
 	}
 
 	values := make([]float64, len(points))
@@ -40,10 +40,10 @@ func TestClosedCompute(t *testing.T) {
 func TestClosedIntegrate(t *testing.T) {
 	basis := NewClosed(1)
 
-	levels := []uint32{0, 1, 2, 3}
-	values := []float64{1, 0.25, 1.0 / 2 / 2, 1.0 / 2 / 2 / 2}
+	levels := []uint64{0, 1, 2, 3}
+	values := []float64{1.0, 0.25, 1.0 / 2 / 2, 1.0 / 2 / 2 / 2}
 
 	for i := range levels {
-		assert.Equal(basis.Integrate([]uint64{compose(levels[i], uint32(0))}), values[i], t)
+		assert.Equal(basis.Integrate([]uint64{compose(levels[i], 0)}), values[i], t)
 	}
 }
